refactor(core): grid cleanup
